Adding a short video about the project
Guys, I'd like to congratulate you for the fantastic job that you have all done with this project! 
Here's a link - https://www.youtube.com/watch?v=LzcbmAmxs-k or just look it up on YouTube as "GinFizz Team Project Bubble Trouble", 
to the video for the project that I made today and hope that you'll like. I was unable to add it to GitHub as the file is too big, but I hope it will be helpful for the presentation tomorrow. Good luck to us all and - Break a leg!